Move common hooks test code to util
